tests/wpt/web-platform-tests/css/css-transforms/animation/translate-interpolation.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>translate interpolation</title>
    <link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
    <link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-translate">
    <meta name="assert" content="translate supports <length> and <percentage> animation.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="resources/interpolation-testcommon.js"></script>
    <style>
      body {
        width: 500px;
        height: 500px;
      }
      div {
        width: 10px;
        height: 10px;
      }
    </style>
  </head>
  <body>
    <script>
      test_interpolation({
        property: 'translate',
        from: '220px 240px 260px',
        to: '300px 400px 500px',
      }, [
        {at: -1, expect: '140px 80px 20px'},
        {at: 0, expect: '220px 240px 260px'},
        {at: 0.125, expect: '230px 260px 290px'},
        {at: 0.875, expect: '290px 380px 470px'},
        {at: 1, expect: '300px 400px 500px'},
        {at: 2, expect: '380px 560px 740px'}
      ]);

      test_interpolation({
        property: 'translate',
        from: '480px 400px 320px',
        to: '240% 160%',
      }, [
        {at: -1, expect: 'calc(960px - 240%) calc(800px - 160%) 640px'},
        {at: 0, expect: '480px 400px 320px'},
        {at: 0.125, expect: 'calc(420px + 30%) calc(350px + 20%) 280px'},
        {at: 0.875, expect: 'calc(210% + 60px) calc(140% + 50px) 40px'},
        {at: 1, expect: '240% 160%'},
        {at: 2, expect: 'calc(480% - 480px) calc(320% - 400px) -320px'}
      ]);

      test_interpolation({
        property: 'translate',
        from: 'none',
        to: '8px 80% 800px',
      }, [
        {at: -1, expect: '-8px -80% -800px'},
        {at: 0, expect: 'none'},
        {at: 0.125, expect: '1px 10% 100px'},
        {at: 0.875, expect: '7px 70% 700px'},
        {at: 1, expect: '8px 80% 800px'},
        {at: 2, expect: '16px 160% 1600px'}
      ]);

      test_interpolation({
        property: 'translate',
        from: 'none',
        to: 'none',
      }, [
        {at: -1, expect: 'none'},
        {at: 0, expect: 'none'},
        {at: 0.125, expect: 'none'},
        {at: 0.875, expect: 'none'},
        {at: 1, expect: 'none'},
        {at: 2, expect: 'none'}
      ]);
    </script>
  </body>
</html>
